repaired a bug that was distoring phylogeny posteriors (when using -j0).  Also several other small modifications

+++ output.cpp
@@ -319,7 +319,7 @@ char* outputbanner(const char *bannertext)
   char holdbanner[BANNERMAXLENGTH];
   strcpy(holdbanner,bannertext);
   char *b;
-  l = strlen(holdbanner);
+  l = (int) strlen(holdbanner);
   assert (l < BANNERMAXLENGTH - 2);
   b = bannerall;
   *b = 0;
